requirements.txt
flask
flask-cors
pymongo
requests
pillow
imagehash
beautifulsoup4
selenium
scikit-image
opencv-python
werkzeug
fastapi
httpx
supabase
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
email-validator
pydantic[email]
